Update with semester gap functionality
Added comments throughout the function. Also added additionally functionality to limit the max number of semesters between courses. The code for limiting the semester gap is very unoptimized, and results in the function taking about 10 times as long for small datasets. If the variable semsBetweenClassesLimit is set to a negative number, the function does not limit the semester gap and runs data on all eligible classes. Also added the makeSemesterNumberColumn function to be called internally by getCorrelationsWithMinNSharedStudents